Refactor Completed - Major Logic Overhaul - Data Storage and Management Introduced
User data is now being stored intelligently, with not waste of memory space, and updated throughout the use of the app. This required a logic overhaul from scratch, since the global variable that tracks user data had to be inserted and manipulated in pretty much every function

Exercise expansion is now absolutely, fully functional. The main components (sets, reps, weight) are already at 100%. More flexibility to be added regarding the additional features, though they are already practically functional as of now.

App is already useable, yet not fool-proof. More UX and visual features to be added.